+++ packages/client/src/components/Input/index.module.scss
@@ -19,12 +19,6 @@
   padding: 5px;
   color: var(--black);
 
-  &::placeholder {
-    color: white;
-    -webkit-text-stroke: 1px var(--grey-80);
-    text-stroke: 1px var(--grey-80);
-  }
-
   &:hover,
   &:focus {
     border-bottom-color: var(--sky);
@@ -47,7 +41,7 @@
 .textarea {
   border-radius: 10px;
   border: 2px solid var(--black);
-  padding: 17px;
+  padding: 16px;
   width: 100%;
   resize: none;
   max-height: 150px;
@@ -56,4 +50,12 @@
   &:focus {
     border-color: var(--sky);
   }
+}
+
+.input::placeholder,
+.textarea::placeholder {
+  color: white;
+  font-size: var(--font-size-20);
+  -webkit-text-stroke: 1px var(--grey-80);
+  text-stroke: 1px var(--grey-80);
 }
