+++ packages/client/src/components/Input/lib/useTextarea.tsx
@@ -8,7 +8,7 @@ type UseTextareaProps = {
 export const useTextarea = ({ textareaRef, value }: UseTextareaProps) => {
 	const resizeTextArea = useCallback(() => {
 		if (textareaRef?.current) {
-			textareaRef.current.style.height = 'min-content';
+			textareaRef.current.style.height = 'inherit';
 			textareaRef.current.style.height = textareaRef.current.scrollHeight + 'px';
 		}
 	}, []);
